Load samba-3.5.0rc1 into branches/samba/upstream-3.5.
git-svn-id: svn://svn.debian.org/svn/pkg-samba/branches/samba/upstream-3.5@3213 fc4039ab-9d04-0410-8cac-899223bdd6b0

+++ lib/talloc/script/release-script.sh
@@ -10,10 +10,20 @@ if [ ! -d "lib/talloc" ]; then
     exit 1
 fi
 
+# Check exports and signatures are up to date
+pushd lib/talloc
+./script/abi_checks.sh talloc talloc.h
+abicheck=$?
+popd
+if [ ! "$abicheck" = "0" ]; then
+    echo "ERROR: ABI Checks produced warnings!"
+    exit 1
+fi
+
 git clean -f -x -d lib/talloc
 git clean -f -x -d lib/replace
 
-curbranch=`git-branch |grep "^*" | tr -d "* "`
+curbranch=`git branch |grep "^*" | tr -d "* "`
 
 version=$1
 strver=`echo ${version} | tr "." "-"`
